improved positioner
