+++ src/popover/stories/index.stories.js
@@ -3,6 +3,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Box from 'ui-box'
 import { Popover } from '../../popover'
+import { TextInputField } from '../../text-input'
 import { Pane } from '../../layers'
 import { Text } from '../../typography'
 import { Button } from '../../buttons'
@@ -10,8 +11,8 @@ import { Position } from '../../positioner'
 
 const PopoverContent = () => (
   <Pane
-    width="240px"
-    height="240px"
+    width={240}
+    height={240}
     display="flex"
     alignItems="center"
     justifyContent="center"
@@ -21,10 +22,23 @@ const PopoverContent = () => (
   </Pane>
 )
 
+const PopoverContentWithTextInput = () => (
+  <Pane
+    width={320}
+    height={200}
+    display="flex"
+    alignItems="center"
+    justifyContent="center"
+    flexDirection="column"
+  >
+    <TextInputField label="Auto focus" autoFocus width="80%" />
+  </Pane>
+)
+
 const ClosablePopoverContent = ({ close }) => (
   <Pane
-    width="240px"
-    height="240px"
+    width={240}
+    height={240}
     display="flex"
     alignItems="center"
     justifyContent="center"
@@ -45,64 +59,26 @@ const controlUsage = (
     content={({ close }) => <ClosablePopoverContent close={close} />}
     display="inline-block"
   >
-    {({ isOpen, toggle, getRef }) => (
+    {({ isShown, toggle, getRef }) => (
       <Button
-        // You can use `isOpen` to set a properties
+        // You can use `isShown` to set a properties
         // Use with caution, calculations are based on the width
         // as soon as you toggle it
-        isActive={isOpen}
+        isActive={isShown}
         // Use toggle to show/hide the popover
         onClick={toggle}
         // GetRef is used to get the ref of the element we need to run
         // getBoundingClientRect() on
         innerRef={ref => getRef(ref)}
       >
-        {isOpen ? 'is open' : 'open'}
+        {isShown ? 'is open' : 'open'}
       </Button>
     )}
   </Popover>
 )
 
 storiesOf('popover', module)
-  .add('Popover', () => (
-    <Box padding={40}>
-      {(() => {
-        document.body.style.margin = '0'
-        document.body.style.height = '100vh'
-      })()}
-      <Popover content={<PopoverContent />}>
-        <Button marginRight={20}>Trigger Popover</Button>
-      </Popover>
-      <Popover
-        content={({ close }) => <ClosablePopoverContent close={close} />}
-      >
-        <Button marginRight={20}>Trigger Closable Popover</Button>
-      </Popover>
-      <Popover
-        useSmartPositioning={false}
-        content={({ close }) => <ClosablePopoverContent close={close} />}
-      >
-        <Button marginRight={20}>Disable Smart Positioning</Button>
-      </Popover>
-      <Pane overflowY="scroll" height={400} appearance="tint3">
-        <Box height={800} paddingTop={200}>
-          <Popover
-            position={Position.BOTTOM_LEFT}
-            content={({ close }) => <ClosablePopoverContent close={close} />}
-          >
-            <Button marginRight={20}>Inside Scrolling Container</Button>
-          </Popover>
-        </Box>
-      </Pane>
-      {controlUsage}
-      <Popover content={<PopoverContent />}>
-        <Button position="absolute" left={40} bottom={40} marginRight={20}>
-          Use Smart Positioning
-        </Button>
-      </Popover>
-    </Box>
-  ))
-  .add('Positioning Sides', () => (
+  .add('positions', () => (
     <Box
       display="flex"
       alignItems="center"
@@ -140,5 +116,54 @@ storiesOf('popover', module)
           </Popover>
         </Box>
       </Box>
+    </Box>
+  ))
+  .add('usages', () => (
+    <Box padding={40}>
+      {(() => {
+        document.body.style.margin = '0'
+        document.body.style.height = '100vh'
+      })()}
+      <Popover content={<PopoverContent />}>
+        <Button marginRight={20}>Trigger Popover</Button>
+      </Popover>
+      <Popover
+        content={({ close }) => <ClosablePopoverContent close={close} />}
+      >
+        <Button marginRight={20}>Trigger Closable Popover</Button>
+      </Popover>
+      <Popover
+        useSmartPositioning={false}
+        content={({ close }) => <ClosablePopoverContent close={close} />}
+      >
+        <Button marginRight={20}>Disable Smart Positioning</Button>
+      </Popover>
+      <Pane overflowY="scroll" height={400} appearance="tint3">
+        <Box height={800} paddingTop={200}>
+          <Popover
+            position={Position.BOTTOM_LEFT}
+            content={({ close }) => <ClosablePopoverContent close={close} />}
+          >
+            <Button marginRight={20}>Inside Scrolling Container</Button>
+          </Popover>
+        </Box>
+      </Pane>
+      {controlUsage}
+      <Popover content={<PopoverContent />}>
+        <Button position="absolute" left={40} bottom={40} marginRight={20}>
+          Use Smart Positioning
+        </Button>
+      </Popover>
+    </Box>
+  ))
+  .add('auto focus text input', () => (
+    <Box padding={120}>
+      {(() => {
+        document.body.style.margin = '0'
+        document.body.style.height = '100vh'
+      })()}
+      <Popover content={<PopoverContentWithTextInput />}>
+        <Button marginRight={20}>Trigger Popover</Button>
+      </Popover>
     </Box>
   ))
