+++ src/positioner/index.js
@@ -1,3 +1,2 @@
 export Positioner from './src/Positioner'
 export Position from './src/Position'
-export Positioner2 from './src/Positioner2'
